Fix compilation for Visual Studio 2017

+++ src/lib/app/Application.cpp
@@ -440,7 +440,7 @@ void Application::updateRecentProjects(const FilePath& projectSettingsFilePath)
 		}
 
 		recentProjects.insert(recentProjects.begin(), projectSettingsFilePath);
-		while (recentProjects.size() > appSettings->getMaxRecentProjectsCount())
+		while (static_cast<int>(recentProjects.size()) > appSettings->getMaxRecentProjectsCount())
 		{
 			recentProjects.pop_back();
 		}
